# swagger/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: Snoopy API
  description: |
    This is a sample Petstore server.  You can find
    out more about Swagger at
    [http://swagger.io](http://swagger.io) or on
    [irc.freenode.net, #swagger](http://swagger.io/irc/).
  contact:
    email: bekhzod.tillakhanov@gmail.com
paths:
  /api/v1/sign-in/:
    post:
      summary: Create new token for authorization.
      operationId: createToken
      tags:
        - Authentication
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                $ref: "schemas/SignInResponseDTO.yaml"
        "400":
          description: Fail
          content:
            application/json:
              schema:
                $ref: "schemas/ErrorDTO.yaml"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "schemas/SignInRequestDTO.yaml"
  /api/v1/sign-out/:
    post:
      summary: Destroy authorization token.
      tags:
        - Authentication
      operationId: destoryToken
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                $ref: "schemas/SignOutResponseDTO.yaml"
        "400":
          description: Fail
          content:
            application/json:
              schema:
                $ref: "schemas/ErrorDTO.yaml"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "schemas/SignOutRequestDTO.yaml"
  /api/v1/forgot-password/:
    post:
      summary: Recovery person password.
      operationId: forgotPassword
      tags:
        - Authentication
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                $ref: "schemas/ForgotPasswordResponseDTO.yaml"
        "400":
          description: Fail
          content:
            application/json:
              schema:
                $ref: "schemas/ErrorDTO.yaml"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "schemas/ForgotPasswordRequestDTO.yaml"
  /api/v1/forgot-password/confirm/:
    post:
      summary: Confirm forgot password.
      operationId: forgotPasswordConfirm
      tags:
        - Authentication
      responses:
        "200":
          description: Successful
          content:
            application/json:
              schema:
                $ref: "schemas/ForgotPasswordConfirmResponseDTO.yaml"
        "400":
          description: Fail
          content:
            application/json:
              schema:
                $ref: "schemas/ErrorDTO.yaml"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "schemas/ForgotPasswordConfirmRequestDTO.yaml"
components:
  securitySchemes:
    Token:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - Token: []